requirements.txt
asgiref==3.6.0
attrs==22.2.0
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.1
drf-spectacular-sidecar==2023.4.1
inflection==0.5.1
jsonschema==4.17.3
PyJWT==2.6.0
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.3
tzdata==2023.3
uritemplate==4.1.1
django-cors-headers==3.14.0
pycpfcnpj==1.7.0